# openapi.yaml
components:
  schemas:
    Dataset:
      additionalProperties: false
      description: A model describing a dataset.
      properties:
        description:
          anyOf:
          - type: string
          - type: 'null'
          description: The description of the dataset.
          title: Description
        files:
          description: Files contained in the dataset.
          items:
            $ref: '#/components/schemas/DatasetFile'
          title: Files
          type: array
        id:
          description: ID of the dataset
          title: Id
          type: string
        stage:
          $ref: '#/components/schemas/WorkType'
          description: Current stage of this dataset.
        title:
          description: The title of the dataset.
          title: Title
          type: string
      required:
      - id
      - stage
      - title
      - description
      - files
      title: Dataset
      type: object
    DatasetFile:
      additionalProperties: false
      description: A file as that is part of a dataset.
      properties:
        extension:
          description: The file extension with a leading dot.
          title: Extension
          type: string
        id:
          description: The file ID.
          title: Id
          type: string
      required:
      - id
      - extension
      title: DatasetFile
      type: object
    WorkPackageCreationData:
      additionalProperties: false
      description: All data necessary to create a work package.
      properties:
        dataset_id:
          title: Dataset Id
          type: string
        file_ids:
          anyOf:
          - items:
              type: string
            type: array
          - type: 'null'
          description: IDs of all included files. If None, all files of the dataset
            are assumed as target.
          title: File Ids
        type:
          $ref: '#/components/schemas/WorkType'
        user_public_crypt4gh_key:
          description: The user's public Crypt4GH key in base64 encoding
          title: User Public Crypt4Gh Key
          type: string
      required:
      - dataset_id
      - type
      - user_public_crypt4gh_key
      title: WorkPackageCreationData
      type: object
    WorkPackageCreationResponse:
      description: Response when a work package has been created.
      properties:
        id:
          description: ID of the work package
          title: Id
          type: string
        token:
          description: The work package access token, encrypted with the user's public
            Crypt4GH key
          title: Token
          type: string
      required:
      - id
      - token
      title: WorkPackageCreationResponse
      type: object
    WorkPackageDetails:
      description: Details about the work package that can be requested.
      properties:
        created:
          description: Creation date of the work package
          format: date-time
          title: Created
          type: string
        expires:
          format: date-time
          title: Expiration date of the work package
          type: string
        files:
          additionalProperties:
            type: string
          description: IDs of all included files mapped to their file extensions
          examples:
          - file-id-1: .json
            file-id-2: .csv
          title: Files
          type: object
        type:
          $ref: '#/components/schemas/WorkType'
      required:
      - type
      - files
      - created
      - expires
      title: WorkPackageDetails
      type: object
    WorkType:
      description: The type of work that a work package describes.
      enum:
      - download
      - upload
      title: WorkType
      type: string
  securitySchemes:
    HTTPBearer:
      scheme: bearer
      type: http
info:
  description: A service managing work packages for the GHGA CLI
  title: Work Package Service
  version: 4.0.1
openapi: 3.1.0
paths:
  /health:
    get:
      description: Used to test if this service is alive
      operationId: health_health_get
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
      summary: health
      tags:
      - WorkPackages
  /users/{user_id}/datasets:
    get:
      description: Endpoint used to get details for all datasets that are accessible
        to the given user
      operationId: get_datasets
      parameters:
      - in: path
        name: user_id
        required: true
        schema:
          title: User Id
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/Dataset'
                title: Response 200 Get Datasets
                type: array
          description: Datasets have been fetched.
        '403':
          description: Not authorized to get datasets.
        '422':
          description: Validation error in submitted user data.
      security:
      - HTTPBearer: []
      summary: Get all datasets of the given user
      tags:
      - Datasets
  /work-packages:
    post:
      description: Endpoint used to create a new work package
      operationId: create_work_package
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WorkPackageCreationData'
        required: true
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkPackageCreationResponse'
          description: Work package was successfully created.
        '403':
          description: Not authorized to create a work package.
        '422':
          description: Validation error in submitted user data.
      security:
      - HTTPBearer: []
      summary: Create a work package
      tags:
      - WorkPackages
  /work-packages/{work_package_id}:
    get:
      description: Endpoint used to get work package details
      operationId: get_work_package
      parameters:
      - in: path
        name: work_package_id
        required: true
        schema:
          title: Work Package Id
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkPackageDetails'
          description: Work package details have been found.
        '403':
          description: Not authorized to get the work package.
        '422':
          description: Validation error in submitted user data.
      security:
      - HTTPBearer: []
      summary: Get a work package
      tags:
      - WorkPackages
  /work-packages/{work_package_id}/files/{file_id}/work-order-tokens:
    post:
      description: Endpoint used to create a work order token
      operationId: create_work_order_token
      parameters:
      - in: path
        name: work_package_id
        required: true
        schema:
          title: Work Package Id
          type: string
      - in: path
        name: file_id
        required: true
        schema:
          title: File Id
          type: string
      responses:
        '201':
          content:
            application/json:
              schema: {}
          description: Work order token has been created.
        '403':
          description: Not authorized to create the work order token.
        '422':
          description: Validation error in submitted user data.
      security:
      - HTTPBearer: []
      summary: Create a work order token
      tags:
      - WorkPackages
servers:
- url: ''
tags:
- name: WorkPackages
- name: Datasets